做完声压测量

+++ app/src/main/java/edu/scut/acoustics/utils/DSPMath.java
@@ -50,5 +50,5 @@ public class DSPMath {
      * @param ff 频率
      * @return 总分压
      */
-    public native float slmfunc(float[] x, float[] l8, float[] ff);
+    public native float slmfunc(short[] x, float[] l8, float[] ff);
 }
